requirements.txt
anthropic==0.39.0
python-dotenv==1.0.1
requests==2.32.3
